api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
elevenlabs>=0.2.27
python-multipart>=0.0.6
sse-starlette>=1.8.2
python-dotenv>=1.0.0
